+++ src/modules/cpuusage/option.h
@@ -4,15 +4,10 @@
 
 #include "common/option.h"
 
-typedef enum FFCPUUsageDisplayType {
-    FF_CPUUSAGE_DISPLAY_TYPE_DEFAULT,
-    FF_CPUUSAGE_DISPLAY_TYPE_SEPARATE,
-} FFCPUUsageDisplayType;
-
 typedef struct FFCPUUsageOptions
 {
     FFModuleBaseInfo moduleInfo;
     FFModuleArgs moduleArgs;
 
-    FFCPUUsageDisplayType displayType;
+    bool separate;
 } FFCPUUsageOptions;